backend/requirements.txt
Flask
Flask_SQLAlchemy
SQLAlchemy
openai
python-dotenv
Flask-Bcrypt
